# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GachaPull - Pokemon & One Piece Gacha API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gacha/pull:
    post:
      operationId: gachaPull
      tags: [gacha]
      summary: Pull from a pack
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GachaPullBody"
      responses:
        "200":
          description: Pull result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GachaPullResult"
        "400":
          description: Insufficient coins or pack unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserListResponse"
  /admin/users/{userId}:
    get:
      operationId: getAdminUser
      tags: [admin]
      summary: Get user detail (admin)
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
    put:
      operationId: updateAdminUser
      tags: [admin]
      summary: Update user (admin)
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/transactions:
    get:
      operationId: listAdminTransactions
      tags: [admin]
      summary: List all payment transactions (admin)
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, completed, failed]
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminTransactionListResponse"
  /admin/settings:
    get:
      operationId: getAdminSettings
      tags: [admin]
      summary: Get payment gateway settings (admin)
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentSettings"
    put:
      operationId: updateAdminSettings
      tags: [admin]
      summary: Update payment gateway settings (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentSettings"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/coin-packages:
    get:
      operationId: listAdminCoinPackages
      tags: [admin]
      summary: List coin packages (admin)
      responses:
        "200":
          description: Coin packages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CoinPackage"
    post:
      operationId: createAdminCoinPackage
      tags: [admin]
      summary: Create coin package (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCoinPackageBody"
      responses:
        "201":
          description: Created coin package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoinPackage"
  /admin/coin-packages/{packageId}:
    put:
      operationId: updateAdminCoinPackage
      tags: [admin]
      summary: Update coin package (admin)
      parameters:
        - in: path
          name: packageId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCoinPackageBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoinPackage"
    delete:
      operationId: deleteAdminCoinPackage
      tags: [admin]
      summary: Delete coin package (admin)
      parameters:
        - in: path
          name: packageId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # Auth schemas
...
    Card:
      type: object
      required: [id, name, franchise, rarity, imageUrl, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        franchise:
          type: string
          enum: [pokemon, onepiece]
        rarity:
          type: string
          enum: [common, rare, super_rare, ultra_rare, legendary]
        imageUrl:
          type: string
        description:
          type: string
          nullable: true
        pullCount:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    GachaPullBody:
      type: object
      required: [packId, pullCount]
      properties:
        packId:
          type: integer
        pullCount:
          type: integer
          enum: [1, 10]
          description: Pull 1 or 10 cards
    GachaPullResult:
      type: object
      required: [cards, coinsSpent, coinsRemaining]
      properties:
        cards:
          type: array
          items:
            $ref: "#/components/schemas/PulledCard"
        coinsSpent:
          type: integer
        coinsRemaining:
          type: integer
    PulledCard:
      type: object
      required: [card, isNew]
      properties:
        card:
          $ref: "#/components/schemas/Card"
        isNew:
          type: boolean
          description: True if this is the first time user got this card
...
    GachaPullRecord:
      type: object
      required: [id, pack, card, pulledAt]
      properties:
        id:
          type: integer
        pack:
          type: object
          required: [id, name]
          properties:
            id:
              type: integer
            name:
              type: string
        card:
          $ref: "#/components/schemas/Card"
        pulledAt:
          type: string
          format: date-time

    # Collection schemas
...
    CoinPackage:
      type: object
      required: [id, name, coins, priceUsd, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        coins:
          type: integer
        priceUsd:
          type: number
        bonusCoins:
          type: integer
          nullable: true
        isPopular:
          type: boolean
          nullable: true
        isActive:
          type: boolean
    CreateCoinPackageBody:
      type: object
      required: [name, coins, priceUsd]
      properties:
        name:
          type: string
        coins:
          type: integer
        priceUsd:
          type: number
        bonusCoins:
          type: integer
          nullable: true
        isPopular:
          type: boolean
          nullable: true
        isActive:
          type: boolean
...
    PaymentRecord:
      type: object
      required: [id, method, amountUsd, coinsGranted, status, createdAt]
      properties:
        id:
          type: integer
        method:
          type: string
          enum: [stripe, midtrans, usdt]
        amountUsd:
          type: number
        coinsGranted:
          type: integer
        status:
          type: string
          enum: [pending, completed, failed]
        createdAt:
          type: string
          format: date-time

    # Leaderboard
...
    AdminUserListResponse:
      type: object
      required: [users, total, page, limit]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUserSummary"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    AdminUserSummary:
      type: object
      required: [id, username, email, role, coinsBalance, totalSpentUsd, totalPulls, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin]
        coinsBalance:
          type: integer
        totalSpentUsd:
          type: number
        totalPulls:
          type: integer
        createdAt:
          type: string
          format: date-time
    AdminUserDetail:
      allOf:
        - $ref: "#/components/schemas/AdminUserSummary"
        - type: object
          properties:
            recentPulls:
              type: array
              items:
                $ref: "#/components/schemas/GachaPullRecord"
            paymentHistory:
              type: array
              items:
                $ref: "#/components/schemas/PaymentRecord"
    AdminUpdateUserBody:
      type: object
      properties:
        role:
          type: string
          enum: [user, admin]
        coinsBalance:
          type: integer
    AdminTransactionListResponse:
      type: object
      required: [transactions, total, page, limit]
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/AdminTransactionRecord"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    AdminTransactionRecord:
      type: object
      required: [id, user, method, amountUsd, coinsGranted, status, createdAt]
      properties:
        id:
          type: integer
        user:
          type: object
          required: [id, username]
          properties:
            id:
              type: integer
            username:
              type: string
        method:
          type: string
          enum: [stripe, midtrans, usdt]
        amountUsd:
          type: number
        coinsGranted:
          type: integer
        status:
          type: string
          enum: [pending, completed, failed]
        createdAt:
          type: string
          format: date-time
    PaymentSettings:
      type: object
      properties:
        stripe:
          type: object
          properties:
            publicKey:
              type: string
            secretKey:
              type: string
            webhookSecret:
              type: string
            enabled:
              type: boolean
        midtrans:
          type: object
          properties:
            serverKey:
              type: string
            clientKey:
              type: string
            isProduction:
              type: boolean
            enabled:
              type: boolean
        usdt:
          type: object
          properties:
            walletAddress:
              type: string
            network:
              type: string
              enum: [trc20, erc20, bep20]
            enabled:
              type: boolean
        nowpayments:
          type: object
          properties:
            apiKey:
              type: string
            enabled:
              type: boolean